Teeworlds
Hello,

I was wondering if you're interested in expanding the Teerace project you worked on long time ago. Currently, it seems that there is a small problem with the site crashing, but I have an idea for a project where tw-clanzone.com will be integrated into the Teerace API/token system, where people would log in in-game and track their stats of wins/kills/etc. If anything, I'm willing to compensate, though I'm somewhat unfamiliar with international transactions as I'm in the US.

Cheers,
Pathos